+++ raiden/transfer/secret_registry.py
@@ -1,4 +1,3 @@
-from raiden.transfer.architecture import Event
 from raiden.transfer.channel import get_status
 from raiden.transfer.events import ContractSendSecretReveal
 from raiden.transfer.state import (
@@ -13,14 +12,17 @@ def events_for_onchain_secretreveal(
         channel_state: NettingChannelState,
         secret: typing.Secret,
         expiration: typing.BlockExpiration,
-) -> typing.List[Event]:
+) -> typing.List[ContractSendSecretReveal]:
     events = list()
 
     if not isinstance(secret, typing.T_Secret):
         raise ValueError('secret must be a Secret instance')
 
     if get_status(channel_state) in CHANNEL_STATES_PRIOR_TO_CLOSED + (CHANNEL_STATE_CLOSED, ):
-        reveal_event = ContractSendSecretReveal(expiration, secret)
+        reveal_event = ContractSendSecretReveal(
+            expiration=expiration,
+            secret=secret,
+        )
         events.append(reveal_event)
 
     return events