+++ raiden/utils/cli.py
@@ -16,6 +16,7 @@ from click.formatting import iter_rows, measure_table, wrap_text
 from pytoml import TomlError, load
 from web3.gas_strategies.time_based import fast_gas_price_strategy, medium_gas_price_strategy
 
+from raiden.constants import Environment
 from raiden.exceptions import InvalidAddress
 from raiden.utils import address_checksum_and_decode
 from raiden_contracts.constants import NETWORKNAME_TO_ID
@@ -270,6 +271,14 @@ class NetworkChoiceType(click.Choice):
         else:
             network_name = super().convert(value, param, ctx)
             return NETWORKNAME_TO_ID[network_name]
+
+
+class EnvironmentChoiceType(click.Choice):
+    def convert(self, value, param, ctx):
+        try:
+            return Environment(value)
+        except ValueError:
+            self.fail(f"'{value}' is not a valid environment type", param, ctx)
 
 
 class GasPriceChoiceType(click.Choice):